First rename of directory structure for SpectatorViewPlugin -> SpectatorView.OpenCV
