+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/CalibrationBoard.png
