+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/QRCodes/sv11.png
