+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/QRCodes/sv14.png
