+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/QRCodes/sv17.png
