+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/QRCodes/sv3.png
