+++ src/SpectatorView.Native/SpectatorView.OpenCV/CalibrationBoardGenerator/QRCodes/sv6.png
