+++ src/SpectatorView.Native/SpectatorView.OpenCV/UWP/SpectatorViewPlugin.h
