+++ src/SpectatorView.Native/SpectatorView.OpenCV/UWP/dllmain.cpp
@@ -1,0 +1,14 @@
+﻿#include "..\SharedFiles\pch.h"
+
+BOOL APIENTRY DllMain(HMODULE /* hModule */, DWORD ul_reason_for_call, LPVOID /* lpReserved */)
+{
+    switch (ul_reason_for_call)
+    {
+    case DLL_PROCESS_ATTACH:
+    case DLL_THREAD_ATTACH:
+    case DLL_THREAD_DETACH:
+    case DLL_PROCESS_DETACH:
+        break;
+    }
+    return TRUE;
+}
