--- src/SpectatorView.Native/SpectatorViewPlugin/SpectatorViewPlugin/CalibrationBoardGenerator/CalibrationBoardGenerator.vcxproj.filters
@@ -1,84 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;ipp;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="CalibrationBoardGenerator.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <Image Include="QRCodes\sv0.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv1.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv2.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv3.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv4.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv5.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv6.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv7.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv8.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv9.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv10.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv11.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv12.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv13.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv14.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv15.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv16.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv17.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv18.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-    <Image Include="QRCodes\sv19.png">
-      <Filter>Resource Files</Filter>
-    </Image>
-  </ItemGroup>
-</Project>
